+++ lib/src/line_splitting/rule_set.dart
@@ -25,12 +25,12 @@ class RuleSet {
   /// Returns `true` of [rule] is bound in this set.
   bool contains(Rule rule) => _values[rule.index] != null;
 
-  /// Gets the bound value for [rule] or `0` if it is not bound.
+  /// Gets the bound value for [rule] or [Rule.unsplit] if it is not bound.
   int getValue(Rule rule) {
     var value = _values[rule.index];
     if (value != null) return value;
 
-    return 0;
+    return Rule.unsplit;
   }
 
   /// Invokes [callback] for each rule in [rules] with the rule's value, which
@@ -66,7 +66,7 @@ class RuleSet {
       if (otherValue == null) {
         // The other rule is unbound, so see if we can constrain it eagerly to
         // a value now.
-        if (constraint == -1) {
+        if (constraint == Rule.mustSplit) {
           // If we know the rule has to split and there's only one way it can,
           // just bind that.
           if (other.numValues == 2) {
@@ -82,16 +82,16 @@ class RuleSet {
       } else {
         // It's already bound, so see if the new rule's constraint disallows
         // that value.
-        if (constraint == -1) {
-          if (otherValue == 0) return false;
+        if (constraint == Rule.mustSplit) {
+          if (otherValue == Rule.unsplit) return false;
         } else if (constraint != null) {
           if (otherValue != constraint) return false;
         }
 
         // See if the other rule's constraint allows us to use this value.
         constraint = other.constrain(otherValue, rule);
-        if (constraint == -1) {
-          if (value == 0) return false;
+        if (constraint == Rule.mustSplit) {
+          if (value == Rule.unsplit) return false;
         } else if (constraint != null) {
           if (value != constraint) return false;
         }