+++ lib/src/rule/type_argument.dart
@@ -43,7 +43,7 @@ class TypeArgumentRule extends Rule {
 
   bool isSplit(int value, Chunk chunk) {
     // Don't split at all.
-    if (value == 0) return false;
+    if (value == Rule.unsplit) return false;
 
     // Split before every argument.
     if (value == numValues - 1) return true;